Subiendo los archivos de configuracion para produccion

+++ database/migrations/2025_04_07_155454_create_inmuebles_table.php
@@ -20,9 +20,9 @@ return new class extends Migration
             $table->string('region', 100);
             $table->string('ciudad', 100);
             $table->string('medida', 100);
-            $table->double('precio', 15, 2);
+            $table->decimal('precio',20);
             $table->double('porcentaje_descuento', 5, 2)->nullable();
-            $table->double('precio_descuento', 15, 2)->nullable();
+            $table->decimal('precio_descuento',20)->nullable();
             $table->string('habitaciones', 5);
             $table->foreignId('id_usuario')->constrained("users");
             $table->foreignId('id_genero')->constrained("generos","id_genero");
